+++ spec/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/QueryGenerator/OptionDeletedQueryGeneratorSpec.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace spec\Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\QueryGenerator;
+
+use PhpSpec\ObjectBehavior;
+use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\NamingUtility;
+use Pim\Bundle\CatalogBundle\Entity\AttributeOption;
+use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
+use Prophecy\Argument;
+
+class OptionDeletedQueryGeneratorSpec extends ObjectBehavior
+{
+    function let(NamingUtility $namingUtility)
+    {
+        $this->beConstructedWith($namingUtility, 'Pim\Bundle\CatalogBundle\Model\AttributeOption', 'code');
+    }
+
+    function it_generates_a_query_to_update_product_select_attributes($namingUtility, AttributeOption $blue, AbstractAttribute $color)
+    {
+        $blue->getAttribute()->willReturn($color);
+        $namingUtility->getAttributeNormFields($color)->willReturn(['normalizedData.color-fr_FR', 'normalizedData.color-en_US']);
+
+        $blue->getCode()->willReturn('blue');
+        $this->generateQuery($blue, 'code', '', '')->shouldReturn([
+            [
+                ['normalizedData.color-fr_FR.code' => 'blue'],
+                ['$unset' => ['normalizedData.color-fr_FR' => '']],
+                ['multiple' => true]
+            ],
+            [
+                ['normalizedData.color-en_US.code' => 'blue'],
+                ['$unset' => ['normalizedData.color-en_US' => '']],
+                ['multiple' => true]
+            ]
+        ]);
+    }
+}