+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/ProductRepository.php
@@ -623,4 +623,43 @@ class ProductRepository extends DocumentRepository implements
 
         return $this;
     }
+
+    /**
+     * {@inheritdoc}
+     *
+     * TODO: find a way to do it efficiently
+     */
+    public function findByProductIdAndOwnerIds($productId, array $ownerIds)
+    {
+        $product = $this->find($productId);
+
+        $ownerIds = array_map(
+            function ($id) {
+                return new \MongoId($id);
+            },
+            $ownerIds
+        );
+
+        // retrieve products whom associations are concerned
+        $qb = $this->createQueryBuilder('p');
+        $qb
+            ->select('associations')
+            ->field('_id')->in($ownerIds)
+            ->field('associations.products.$id')->equals(new \MongoId($productId))
+        ;
+
+        $products = $qb->getQuery()->execute();
+        $associations = [];
+
+        // filter associations
+        foreach ($products as $dummyProduct) {
+            foreach ($dummyProduct->getAssociations() as $association) {
+                if ($association->hasProduct($product)) {
+                    $associations[] = $association;
+                }
+            }
+        }
+
+        return $associations;
+    }
 }