+++ src/Pim/Bundle/CatalogBundle/Repository/AssociationRepositoryInterface.php
@@ -21,4 +21,14 @@ interface AssociationRepositoryInterface
      * @return mixed
      */
     public function countForAssociationType(AssociationType $associationType);
+
+    /**
+     * Get the list of associations corresponding to the given owner IDs
+     *
+     * @param int   $productId
+     * @param array $ownerIds
+     *
+     * @return \Pim\Bundle\CatalogBundle\Model\Association[]
+     */
+    public function findByProductIdAndOwnerIds($productId, array $ownerIds);
 }